requirements.txt
# Core FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database and ORM
sqlmodel==0.0.14
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# Redis and Background Jobs
redis==5.0.1
celery==5.3.4

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
cryptography==41.0.7

# File Storage and Processing
boto3==1.34.0
minio==7.2.0
wkhtmltopdf==0.2
weasyprint==60.2
reportlab==4.0.7

# HTTP Client and External APIs
httpx==0.25.2
aiohttp==3.9.1

# AI and Machine Learning
openai==1.3.7
google-cloud-speech==2.21.0
google-cloud-aiplatform==1.38.1
anthropic==0.7.8

# Monitoring and Logging
sentry-sdk[fastapi]==1.38.0
structlog==23.2.0

# Environment and Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Rate Limiting and Security
slowapi==0.1.9
python-multipart==0.0.6

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3
jinja2==3.1.2
pillow==10.1.0

# Digital Signatures and PDF Processing
pypdf==3.17.4
qrcode[pil]==7.4.2
pycryptodome==3.19.0
pyopenssl==23.3.0
requests==2.31.0

# Telemedicine dependencies
PyJWT==2.8.0
websockets==12.0